feat: restructure some files
