+++ src/electron/splashScreen.ts
